BREAKING:Move ciw helper function into its own file
